fix: 修复生成默认配置文件的逻辑，确保在文件存在且为空时写入默认内容

+++ backend-src/UZonMailService/Program.cs
@@ -20,7 +20,7 @@ Directory.SetCurrentDirectory(AppContext.BaseDirectory);
 
 // 生成默认的配置文件
 var productConfig = "appsettings.Production.json";
-if (!File.Exists(productConfig))
+if (File.Exists(productConfig) && string.IsNullOrWhiteSpace(File.ReadAllText(productConfig)))
 {
     File.WriteAllText(productConfig, "{\n}");
 }
